Refactor: AddEventModal 스타일 분리
Refs: #175
